requirements.txt
matplotlib>=3.4.3
numpy>=1.21.2
tensorflow>=2.6.0
pandas>=1.3.3
flickrapi>=2.4.0
python-dotenv>=0.19.1
requests>=2.26.0
opencv-python>=4.5.3.56